refactor(polish): extract LLM output sanitization into polish/output_cleaning.rs — behavior unchanged, 446 tests green
Pure move of 9 functions + 2 consts (clean_polish_output, strip_thinking_blocks,
<think>/markdown-fence/boilerplate strippers) and their 4 unit tests out of polish.rs
into a child module. clean_polish_output stays pub(crate) and is re-exported via
'pub(crate) use output_cleaning::*;'. Dropped the now-unused 'use std::borrow::Cow'
in polish.rs (the only Cow user moved with the code). No logic/signature/public-API
changes; polish.rs 3663→3436 lines.

cargo check: 0 errors, 87 warnings (unchanged) · clippy: 0 errors
cargo test --lib: 446 passed · tsc / npm run build: green
